Working stained glass effect
Added StainedGlassFilter class, that is used in the paint method of the
component. It implements BufferedImageOp.
Added the VoronoiCell class, which represents the mathematical Voronoi
cell. It contains a generator point (seed), vector of pixels contained
in the cell and vector of their colors. The latter is used to get the
average color of the whole cell.
1. It still works slow!
2. There is a problem with finding the average color of the cells.